requirements.txt
langchain>=0.0.109
streamlit>=1.18.1
pinecone-client
openai
promptlayer
streamlit_option_menu
